refactor!: reorder for clarity since aux_loss changes little after a step
A step of net loss modifies the `EntropyBottleneck.likelihood` slightly,
but the distributions should stabilize pretty quickly anyways.
One other slight difference is that we now grad clip the
aux_loss-associated parameters too, but that shouldn't matter either.

If perfect aux_loss is desired, there are much more precise and faster
methods to achieve this:
https://github.com/InterDigitalInc/CompressAI/pull/231