fix: correctly measure bpp on strings that are not list[list[bytes]]
This does not break existing code, e.g., the official CompressAI models
in which `strings` is of type `list[list[bytes]]`.

It merely allows more formats of strings to work by recursively
searching `strings` for objects of type `bytes`.

Thus, bpp is now measured correctly on the following as well:

```
strings == [
    [b"this works too now", [b"wow nested", [b"wow super nested"]]],
    {"y": b"compressed data", "z": b"compressed data"},
]
```